+++ Changelog.md
@@ -1,3 +1,8 @@
+## v0.0.4
+
+- feat(scan): saving wappalyzer results to file ~ @djnnvx
+- feat(scan): adding cvemap module ~ @djnnvx
+
 ## v0.0.3
 
 - feat: bumping gobuster to latest version (unstable) ~ @djnnvx